% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_functions.R
\name{balanceBinVar}
\alias{balanceBinVar}
\title{Balance indicators for binary variables}
\usage{
balanceBinVar(data, varBalance, match_id, varGroup, pairGroups)
}
\description{
Balance indicators for binary variables
}
\keyword{internal}
