% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_functions.R
\name{balanceCatVar}
\alias{balanceCatVar}
\title{Balance indicators for categorical variables (more than 2 levels)}
\usage{
balanceCatVar(data, varBalance, match_id, varGroup, pairGroups)
}
\description{
Balance indicators for categorical variables (more than 2 levels)
}
\keyword{internal}
