% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_functions.R
\name{balanceContVar}
\alias{balanceContVar}
\title{Balance indicators for continuous variables}
\usage{
balanceContVar(data, varBalance, match_id, varGroup, pairGroups, varWeights)
}
\description{
Balance indicators for continuous variables
}
\keyword{internal}
