% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{checkData}
\alias{checkData}
\title{Check coherence of inputs with data}
\usage{
checkData(
  formulaMatch,
  start,
  data,
  exactMatch,
  vectorK,
  checkOnePerGroup = TRUE
)
}
\description{
Check coherence of inputs with data
}
\keyword{internal}
