% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{checkInputs}
\alias{checkInputs}
\title{Check that type of inputs is appropriate}
\usage{
checkInputs(
  formulaMatch,
  start,
  data,
  distance,
  exactMatch,
  iterate,
  niter_max,
  verbose,
  vectorK
)
}
\description{
Check that type of inputs is appropriate
}
\keyword{internal}
