% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_functions.R
\name{computePairwiseDistances}
\alias{computePairwiseDistances}
\title{Given a set of units, compute all the pairwise distances.}
\usage{
computePairwiseDistances(dataPolygon, distance, Sigma)
}
\description{
Given a set of units, compute all the pairwise distances.
}
\keyword{internal}
