% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_functions.R
\name{evaluateMatching}
\alias{evaluateMatching}
\title{Compute total distance of a matched sample}
\usage{
evaluateMatching(data, varIndexMatch, varsMatch, distance, Sigma)
}
\description{
Compute total distance of a matched sample
}
\keyword{internal}
