% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_functions.R
\name{pairwiseDistance}
\alias{pairwiseDistance}
\title{Distance between two units}
\usage{
pairwiseDistance(A, B, distance, Sigma)
}
\description{
Distance between two units
}
\keyword{internal}
