% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_functions.R
\name{stdzDistances}
\alias{stdzDistances}
\title{Compute factor to standardize distances when matching multiple subjects in one group}
\usage{
stdzDistances(vectorK, withinGroupDist)
}
\description{
Compute factor to standardize distances when matching multiple subjects in one group
}
\keyword{internal}
