% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_functions.R
\name{weighted.variance}
\alias{weighted.variance}
\title{Function to compute weighted variance}
\usage{
weighted.variance(x, w)
}
\description{
Function to compute weighted variance
}
\keyword{internal}
