% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_functions.R
\name{evaluateMatching}
\alias{evaluateMatching}
\title{Compute total distance of a matched sample}
\usage{
evaluateMatching(
  data,
  varIndexMatch,
  varGroup,
  varsMatch,
  distance,
  Sigma,
  dat_stdzDistances
)
}
\description{
Compute total distance of a matched sample
}
\keyword{internal}
